School Person Records due 02/24/25
School Person Records
In this exercise we create a School Person Records application that in addition to Inheritance and
Polymorphism will use FileIO, ArrayLists and Javadoc documentation.
● Please read the whole assignment and make sure you understand it before you start work.
It is always good practice to understand the overall concept and how different parts
interact with each other before you start implementation.
● Make sure you write all the appropriate constructors, getters and setters in each class.
● Make sure to commit the code to GitHub. This is a graded activity.
1
Create a project named SchoolRecords. Create a driver class called
TestSchoolRecords. The driver class contains main()
2
Create a class named Person. Include these data members.
● Name
● Address
● Phone Number
● Email address
3
Create two new classes called Student and Employee that are subclasses (inherits
from) of Person.
4
Student should have a field named status. The value will be freshman, sophomore,
junior, or senior.
5 Employee should have data members for office, salary and date hired.
6
Create two new classes called Faculty and Staff that are subclasses (inherits from) of
Employee.
7
Faculty should have Office hours and rank (junior, senior) as data members. Office
hours can just be a string.
8 Staff has a title as a data member. (Use any string as a value for this field)
9
Create a class named Database. This class should have a method called writePerson
that takes a Person object as an argument and writes the name of the class for the
object and all the information for the object to a file named PersonDatabase. The
name of the file PersonDatabase should be defined as a constant in the Database
class. The constructor for the Database class should create the file PersonDatabase if
one does not exist.
10 Create a method in Database called readDatabase. This method should read all the
persons from the PersonDatabase file and return an ArrayList of Person.
11
In main(), create 6 persons: 2 students, 2 faculty and 2 staff. Assign values to all the
fields of the objects. You do not need to prompt a user to create these objects. Write
these objects one at a time to the PersonDatabase file using the writePerson() method
in Database.
12
In main(), read all the persons in the PersonDatabase file using the readDatabase()
method and print them to the screen by traversing the ArrayList using an enhanced for
loop and the toString() method for each object.
13 Generate Javadoc documentation for each of the classes. Make sure it has all the
components discussed in class.