Phone Designer Due on Mar 3, 2025 11:59 PM
Instructions
You own a phone design company, Phone4U. Your company will manufacture three different models of phone: FlipModel, FoldModel, and BarModel. The default specifications of each of the phones are as follows:

FlipModel: HorseRacer 300 processor, 3MB cache, 16GB storage

FoldModel: SpeedRacer 800 processor, 6MB cache, 64GB storage

BarModel: DragonSlayer 600 processor, 8MB cache, 32MB storage

Each of the phones also has an IMEI which is modeled as an ArrayList of characters. Each of the phones should be cloneable so that the IMEI on the phone can be changed on the clone without affecting the original phone. 

Each of these phones should be comparable to other phones of the same model and so should implement the Comparable interface. They should only be compared using the cache and storage options.

Define an interface named Repairable with two methods howToRepair() and costToRepair(). The BarModel and FlipModel are repairable and should implement this interface. The howToRepair() method should return a String that describes how to repair the phone. The costToRepair() method should return a dollar amount.

Implement a driver class that will test these features.