Update TestSchoolRecords.java
Updated and reworked according to these notes:
You are creating a new FileWriter every time you write a person to the file, which is wasteful. The FileWriter only needs to be created once.  (-2 pts)

ReadDatabase should return a list of Person, not String. Very much like what we did in class with Geometric Objects.  (-8 pts)

Please fix and resubmit.