Rotations on AVL Trees
Instructions
This assignment is on AVL Trees. In class we wrote a class to model an AVL tree and implemented the code for a LL rotation and an LR rotation. Your task for this homework is to write the code for a RL rotation and a RR rotation. You also need to keep track of the path when inserting a node so that you can update the heights correctly. You should be able to get all the code working, at least for insertion. The way to test your code would be to insert nodes with Keys in increasing order and when you are done making sure that the height of the root node is log n.  Another order of keys you may try is 25, 20, 5, 34, 50, 30, 10. 

Here is a test file that might be useful to test your code: MyAVLTreeTest.java

If the AVL tree you have does not work with the test, please use this one: MyAVLTree.java

The output of this test would be:

25

20 25

5 20 25

After LL rotation at node 25

Node Height Balance Factor LeftChild RightChild

5        0         0                         null            null

20      1         0                         5                25

25      0         0                        null            null

5 20 25 34

5 20 25 34 50

After RR rotation at node 25

Node Height Balance Factor LeftChild RightChild

5         0          0                        null              null

20       2          1                         5                 34

25       0          0                       null               null

34      1           0                      25                 50

50      0          0                      null                null

5 20 25 30 34 50

After RL rotation at node 25

Node Height Balance Factor LeftChild RightChild

5        0            0                      null           null

20      1           -1                      5              null

25      2            0                      20            34

30      0            0                      null           null

34      1            0                     30             50

50      0            0                     null           null

5 10 20 25 30 34 50

After LR rotation at node 25

Node Height Balance Factor LeftChild RightChild

5         0         0                       null          null

10       1         0                       5             20

20       0         0                      null          null

25       2         0                     10             34

30       0         0                     null          null

34       1         0                     30           50

50       0         0                    null           null