DFS on Graph/Exam 4
Create a maven project named Exam4. Download Graph.java and place it in the src folder. Download GraphTest.java and place it in the test folder.

There are multiple TODOs at the bottom of Graph.java. You are to complete the DFS and BFS methods for this assignment. The dfs() method  will return a list of vertices in the depth-first search order. The bfs() method will return a list of vertices in the breadth-first search order.  Make sure to test both methods. The last method findMinimumSpanningTree() is part of Exam 4.